初始化PLAY_ID变量,可能可以解决部分机型来短信播放铃声死机问题

+++ MySMSYS/Mss3/DaemonCSM.h
@@ -43,7 +43,7 @@ private:
   GBSTMR chktmr;
   int temp_tmr_index;
   int is_new_proc;
-  short PLAY_ID;
+  int PLAY_ID;
   Vibra *vibra;
   int AddDlgCsmID(int id);
   void AddDlgCsmID_Forced0(int id);
